+++ src/main/kotlin/yqloss/yqlossclientmixinkt/module/corpsefinder/CorpseFinderOptions.kt
@@ -22,11 +22,14 @@ import yqloss.yqlossclientmixinkt.module.option.YCColor
 import yqloss.yqlossclientmixinkt.module.option.YCModuleOptions
 
 interface CorpseFinderOptions : YCModuleOptions {
-    val showExit: Boolean
-    val exitColor: YCColor
+    val exitColor: YCColor?
     val lapis: CorpseOption
     val umber: CorpseOption
     val tungsten: CorpseOption
     val vanguard: CorpseOption
+    val bowmanColor: YCColor?
+    val caverColor: YCColor?
+    val mageColor: YCColor?
+    val muttColor: YCColor?
     val forceEnabled: Boolean
 }