+++ src/main/kotlin/yqloss/yqlossclientmixinkt/module/corpsefinder/CorpseType.kt
@@ -21,12 +21,12 @@ package yqloss.yqlossclientmixinkt.module.corpsefinder
 enum class CorpseType(
     val id: String,
     val armor: String,
-    val option: CorpseOption,
+    val option: () -> CorpseOption,
 ) {
-    LAPIS("lapis", "LAPIS_ARMOR_", CorpseFinder.options.lapis),
-    UMBER("umber", "ARMOR_OF_YOG_", CorpseFinder.options.umber),
-    TUNGSTEN("tungsten", "MINERAL_", CorpseFinder.options.tungsten),
-    VANGUARD("vanguard", "VANGUARD_", CorpseFinder.options.vanguard),
+    LAPIS("lapis", "LAPIS_ARMOR_", { CorpseFinder.options.lapis }),
+    UMBER("umber", "ARMOR_OF_YOG_", { CorpseFinder.options.umber }),
+    TUNGSTEN("tungsten", "MINERAL_", { CorpseFinder.options.tungsten }),
+    VANGUARD("vanguard", "VANGUARD_", { CorpseFinder.options.vanguard }),
     ;
 
     companion object {